+++ tests/misc/fallback_ast_src/switch.php
@@ -2,3 +2,9 @@
 switch($x) {}
 switch($x) {case 2: default:}
 switch($x) {case 2: $y = 3;}
+switch($x) {
+    case 2:
+        $y = 3;
+    default:
+        echo "test";
+}